Move digital image related code from SolrDocument to a concern, plus code refactor
Rename method to indicate behavior
